chore: migrate the project to the latest version of keycloakify to support all new pages
