+++ src/keycloak-theme/login/KcApp.tsx
@@ -1,11 +1,12 @@
 import { lazy, Suspense } from "react";
 import type { KcContext } from "./kcContext";
 import { useI18n } from "./i18n";
-import Fallback, { defaultKcProps, type KcProps, type PageProps } from "keycloakify";
-import {useDownloadTerms} from "keycloakify/lib/pages/Terms";
-import Template from "keycloakify/lib/Template";
+import Fallback, { type PageProps } from "keycloakify/login";
+import {useDownloadTerms} from "keycloakify/login/lib/useDownloadTerms";
 import tos_en_url from "./assets/tos_en.md";
 import tos_fr_url from "./assets/tos_fr.md";
+
+const DefaultTemplate = lazy(() => import("keycloakify/login/Template"));
 
 // You can uncomment this to see the values passed by the main app before redirecting.  
 //import { foo, bar } from "./valuesTransferredOverUrl";
@@ -13,14 +14,11 @@ import tos_fr_url from "./assets/tos_fr.md";
 
 const Empty = lazy(()=> import("./pages/Empty"));
 
-// This is like editing the theme.properties 
+// This is like adding classes to theme.properties 
 // https://github.com/keycloak/keycloak/blob/11.0.3/themes/src/main/resources/theme/keycloak/login/theme.properties
-const kcProps: KcProps = {
-    ...defaultKcProps,
+const classes: PageProps<any, any>["classes"] = {
     // NOTE: The classes are defined in ./KcApp.css
-    // You can add your classes alongside thoses that are present in the default Keycloak theme...
-    "kcHtmlClass": [...defaultKcProps.kcHtmlClass, "my-root-class"],
-    // ...or overwrite  
+    "kcHtmlClass": "my-root-class",
     "kcHeaderWrapperClass": "my-color my-font"
 };
 
@@ -62,25 +60,14 @@ export default function App(props: { kcContext: KcContext; }) {
     * i18n.msg("foo") === <span>foo in English</span>
     */
 
-    const pageProps: Omit<PageProps<any, typeof i18n>, "kcContext"> = {
-        i18n,
-        // Here we have overloaded the default template, however you could use the default one with:  
-        //Template: DefaultTemplate,
-        Template,
-        // Wether or not we should download the CSS and JS resources that comes with the default Keycloak theme.  
-        doFetchDefaultThemeResources: true,
-        ...kcProps,
-    };
-
     return (
         <Suspense>
             {(() => {
                 switch (kcContext.pageId) {
-                    case "empty.ftl": return <Empty {...{ kcContext, ...pageProps }} />;
-                    default: return <Fallback {...{ kcContext, ...pageProps }} />;
+                    case "empty.ftl": return <Empty {...{ kcContext, i18n, classes }} Template={DefaultTemplate} doUseDefaultCss />;
+                    default: return <Fallback {...{ kcContext, i18n, classes }} Template={DefaultTemplate} doUseDefaultCss />;
                 }
             })()}
         </Suspense>
     );
-
 }
