+++ src/keycloak-theme/login/Template.stories.tsx
@@ -1,6 +1,6 @@
 import type {ComponentMeta} from '@storybook/react';
 import KcApp from './KcApp';
-import {template} from '../../.storybook/util'
+import {template} from '../../../.storybook/util'
 
 const bind = template('empty.ftl');
