+++ src/keycloak-theme/login/pages/IdpReviewUserProfile.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta } from '@storybook/react';
 import KcApp from '../KcApp';
-import { template } from '../../../.storybook/util'
+import { template } from '../../../../.storybook/util'
 
-const bind = template('login-otp.ftl');
+const bind = template('idp-review-user-profile.ftl');
 
 export default {
     kind: 'Page',
-    title: 'Theme/Pages/Login/OTP',
+    title: 'Theme/Pages/IDP/Review User Profile',
     component: KcApp,
     parameters: {
         layout: 'fullscreen',
@@ -14,3 +14,4 @@ export default {
 } as ComponentMeta<typeof KcApp>;
 
 export const Default = bind({})
+
