+++ src/keycloak-theme/login/pages/LoginIdpLinkConfirm.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta } from '@storybook/react';
 import KcApp from '../KcApp';
-import { template } from '../../../.storybook/util'
+import { template } from '../../../../.storybook/util'
 
-const bind = template('login-update-password.ftl');
+const bind = template('login-idp-link-confirm.ftl');
 
 export default {
     kind: 'Page',
-    title: 'Theme/Pages/Actions/Update Password',
+    title: 'Theme/Pages/IDP/Confirm Link',
     component: KcApp,
     parameters: {
         layout: 'fullscreen',
@@ -14,3 +14,4 @@ export default {
 } as ComponentMeta<typeof KcApp>;
 
 export const Default = bind({})
+
