+++ src/keycloak-theme/login/pages/LoginIdpLinkEmail.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta } from '@storybook/react';
 import KcApp from '../KcApp';
-import { template } from '../../../.storybook/util'
+import { template } from '../../../../.storybook/util'
 
-const bind = template('terms.ftl');
+const bind = template('login-idp-link-email.ftl');
 
 export default {
     kind: 'Page',
-    title: 'Theme/Pages/Actions/Terms',
+    title: 'Theme/Pages/IDP/Confirm Email',
     component: KcApp,
     parameters: {
         layout: 'fullscreen',
@@ -14,3 +14,4 @@ export default {
 } as ComponentMeta<typeof KcApp>;
 
 export const Default = bind({})
+
