+++ src/keycloak-theme/login/pages/LoginPassword.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta } from '@storybook/react';
 import KcApp from '../KcApp';
-import { template } from '../../../.storybook/util'
+import { template } from '../../../../.storybook/util'
 
-const bind = template('login-idp-link-confirm.ftl');
+const bind = template('login-password.ftl');
 
 export default {
     kind: 'Page',
-    title: 'Theme/Pages/IDP/Confirm Link',
+    title: 'Theme/Pages/Login/Password',
     component: KcApp,
     parameters: {
         layout: 'fullscreen',
@@ -14,4 +14,3 @@ export default {
 } as ComponentMeta<typeof KcApp>;
 
 export const Default = bind({})
-
