+++ src/keycloak-theme/login/pages/LoginUpdatePassword.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta } from '@storybook/react';
 import KcApp from '../KcApp';
-import { template } from '../../../.storybook/util'
+import { template } from '../../../../.storybook/util'
 
-const bind = template('login-username.ftl');
+const bind = template('login-update-password.ftl');
 
 export default {
     kind: 'Page',
-    title: 'Theme/Pages/Login/Username',
+    title: 'Theme/Pages/Actions/Update Password',
     component: KcApp,
     parameters: {
         layout: 'fullscreen',
@@ -14,4 +14,3 @@ export default {
 } as ComponentMeta<typeof KcApp>;
 
 export const Default = bind({})
-export const WithEmailAsUsername = bind({ realm: { loginWithEmailAllowed: true, registrationEmailAsUsername: true } })
