+++ src/keycloak-theme/login/pages/LoginUsername.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta } from '@storybook/react';
 import KcApp from '../KcApp';
-import { template } from '../../../.storybook/util'
+import { template } from '../../../../.storybook/util'
 
-const bind = template('logout-confirm.ftl');
+const bind = template('login-username.ftl');
 
 export default {
     kind: 'Page',
-    title: 'Theme/Pages/Login/Logout Confirmation',
+    title: 'Theme/Pages/Login/Username',
     component: KcApp,
     parameters: {
         layout: 'fullscreen',
@@ -14,3 +14,4 @@ export default {
 } as ComponentMeta<typeof KcApp>;
 
 export const Default = bind({})
+export const WithEmailAsUsername = bind({ realm: { loginWithEmailAllowed: true, registrationEmailAsUsername: true } })
