+++ src/keycloak-theme/login/pages/Terms.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta } from '@storybook/react';
 import KcApp from '../KcApp';
-import { template } from '../../../.storybook/util'
+import { template } from '../../../../.storybook/util'
 
-const bind = template('idp-review-user-profile.ftl');
+const bind = template('terms.ftl');
 
 export default {
     kind: 'Page',
-    title: 'Theme/Pages/IDP/Review User Profile',
+    title: 'Theme/Pages/Actions/Terms',
     component: KcApp,
     parameters: {
         layout: 'fullscreen',
@@ -14,4 +14,3 @@ export default {
 } as ComponentMeta<typeof KcApp>;
 
 export const Default = bind({})
-
