+++ src/keycloak-theme/login/pages/UpdateUserProfile.stories.tsx
@@ -1,12 +1,12 @@
 import { ComponentMeta } from '@storybook/react';
 import KcApp from '../KcApp';
-import { template } from '../../../.storybook/util'
+import { template } from '../../../../.storybook/util'
 
-const bind = template('login-idp-link-email.ftl');
+const bind = template('update-user-profile.ftl');
 
 export default {
     kind: 'Page',
-    title: 'Theme/Pages/IDP/Confirm Email',
+    title: 'Theme/Pages/Actions/Update User Profile',
     component: KcApp,
     parameters: {
         layout: 'fullscreen',
@@ -14,4 +14,3 @@ export default {
 } as ComponentMeta<typeof KcApp>;
 
 export const Default = bind({})
-
